allow different h5 naming and batch writing of h5 data

+++ run_single_session.py
@@ -1,10 +1,11 @@
 import suite2p
 from suite2p.run_s2p import run_s2p
-import preprocessing as pp
+import s2p_preprocessing as pp
 import os
 import shutil
 
-def sbx_run_pipeline(files, h5_overwrite = False, s2p_overwrite=False,ops_d={},db_d={}):
+def sbx_run_pipeline(files, h5_overwrite = False, s2p_overwrite=False,keep_binary=True,
+                    h5_output=None,ops_d={},db_d={}):
     '''sbx filename stem or list of sbx filename stems
     overwrite h5 file if it exists
     overwrite s2p results
@@ -28,27 +29,28 @@ def sbx_run_pipeline(files, h5_overwrite = False, s2p_overwrite=False,ops_d={},d
 
     # run for a single file
     def single_file(f):
-
+        head,tail = os.path.split(f)
         h5exists = check4h5(f)
         if not h5exists or h5_overwrite:
-            h5fname = pp.sbx2h5(f)
+            h5fname = pp.sbx2h5(f,output_name = os.path.join(h5_output,tail+".h5"))
         else:
-            h5fname = f+'.h5'
+            if h5_output is not None:
+                h5fname = f+'.h5'
+            else:
+                h5fname =  os.path.join(h5_output,tail+".h5")
 
         s2pexists = check4s2p(f)
-        head,tail = os.path.split(f)
+
         if s2pexists:
             if s2p_overwrite:
                 shutil.rmtree(os.path.join(head,'suite2p'))
-                bindir = os.path.join(db['fast_disk'],'suite2p')
-                try:
-                    shutil.rmtree(bindir)
             else:
                 print("%s suite2p already done" % f)
                 return None
 
 
         # set ops
+        ops_d['save_path0']=head
         ops = pp.set_ops(ops_d)
 
         # set db
@@ -63,7 +65,8 @@ def sbx_run_pipeline(files, h5_overwrite = False, s2p_overwrite=False,ops_d={},d
         run_s2p(ops=ops,db=db)
 
         # move registered binary over to save data folders
-        shutil.move(os.path.join(bindir,"plane0","data.bin"),os.path.join(head,"suite2p","data.bin"))
+        if keep_binary:
+            shutil.move(os.path.join(bindir,"plane0","data.bin"),os.path.join(head,"suite2p","data.bin"))
         # delete temporary files
         shutil.rmtree(bindir)
